refactored CLI per new command format
